backennd/requirements.txt
dj-database-url==0.5.0
Django==2.2
django-filter==2.1.0
django-rest-auth==0.9.5
djangorestframework==3.9.2
psycopg2-binary
python-decouple==3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
gunicorn==19.9.0
django-heroku
whitenoise==4.1.2

